// src/base/interfaces/ICommand.ts
// src/base/interfaces/ICommand.ts

import { AutocompleteInteraction, ChatInputCommandInteraction } from "discord.js";
import CustomClient from "../classes/CustomClient";
import Category from "../enums/Category";


export default interface ICommand {
    client: CustomClient
    name: string;
    description: string;
    category: Category
    options: object; 
    default_member_permissions: bigint;
    dm_permission: boolean;
    cooldown: number;

    Execute(Interaction: ChatInputCommandInteraction): void;
    AutoComplete(interaction: AutocompleteInteraction): void;
}