// src/base/interfaces/ICommandOptions.ts
// src/base/interfaces/ICommandOptions.

import Category from "../enums/Category";


export default interface ICommandOptions {
    name: string;
    description: string;
    category: Category
    options: object; 
    default_member_permissions: bigint;
    dm_permission: boolean;
    cooldown: number;

}