// src/base/classes/Command.ts
// src/base/classes/Command.ts

import { AutocompleteInteraction, ChatInputCommandInteraction } from "discord.js";
import Category from "../enums/Category";
import ICommand from "../interfaces/ICommand";
import ICommandOptions from "../interfaces/ICommandOptions";
import CustomClient from "./CustomClient";


export default class Command implements ICommand {
    client: CustomClient
    name: string;
    description: string;
    category: Category
    options: object; 
    default_member_permissions: bigint;
    dm_permission: boolean;
    cooldown: number;

    constructor(client: CustomClient, options: ICommandOptions) {
        this.client = client;
        this.name = options.name;
        this.description = options.description;
        this.category = options.category;
        this.options = options.options;
        this.default_member_permissions = options.default_member_permissions;
        this.dm_permission = options.dm_permission;
        this.cooldown = options.cooldown;
    }

    Execute(Interaction: ChatInputCommandInteraction): void {

    }

    AutoComplete(interaction: AutocompleteInteraction): void {

    }
}